requirements.txt
Django~=4.1.6
environ~=1.0
pip~=22.0.2
attrs~=21.2.0
distro~=1.7.0
wheel~=0.37.1
cryptography~=3.4.8
Pillow~=9.4.0
docutils~=0.17.1
tornado~=6.1
Jinja2~=3.0.3
keyring~=23.5.0
pytz~=2022.7.1
asgiref~=3.6.0
sqlparse~=0.4.3
psycopg2~=2.9.5
colorama~=0.4.4
PyYAML~=5.4.1
numpy~=1.21.5
requests~=2.25.1
setuptools~=59.6.0
packaging~=21.3
pyparsing~=2.4.7
zipp~=1.0.0
future~=0.18.2
djangorestframework~=3.14.0
Markdown~=3.4.1
Pygments~=2.11.2